+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to the "ERCx" extension will be documented in this file.
 
 ## [Unreleased]
 
+- Add the inconclusive tests.
+
 ## [0.0.13] - 2024-02-23
 
 - Add support for ERC721 test suite.